feat: SNS 관련 ViewModel 코드 리팩토링

+++ Tlog/app/src/main/java/com/tlog/data/model/response/course/CoursesDestination.kt
@@ -9,7 +9,7 @@ data class CoursesDestination(
     val city: String,
     val location: Location,
     val reviewCount: Int,
-    val averageRating: Int,
+    val averageRating: Double,
     val imageUrl: String?,
     val description: String,
     val tagCountList: List<TagCount>?
